attributes module
