--- packages/vfx/src/util/registerAttributes.ts
@@ -1,10 +1,0 @@
-import { BufferAttribute, BufferGeometry } from "three"
-
-export function registerAttributes(
-  geometry: BufferGeometry,
-  attributes: Record<string, BufferAttribute>
-) {
-  for (const name in attributes) {
-    geometry.setAttribute(name, attributes[name])
-  }
-}
